+++ src/test/java/org/openlmis/fulfillment/OrderDataBuilder.java
@@ -15,8 +15,8 @@
 
 package org.openlmis.fulfillment;
 
+import static org.openlmis.fulfillment.domain.OrderStatus.FULFILLING;
 import static org.openlmis.fulfillment.domain.OrderStatus.ORDERED;
-import static org.openlmis.fulfillment.domain.OrderStatus.PICKING;
 import static org.openlmis.fulfillment.domain.OrderStatus.SHIPPED;
 import static org.openlmis.fulfillment.domain.OrderStatus.TRANSFER_FAILED;
 
@@ -129,8 +129,8 @@ public class OrderDataBuilder {
     return this;
   }
 
-  public OrderDataBuilder withPickingStatus() {
-    return withStatus(PICKING);
+  public OrderDataBuilder withFulfillingStatus() {
+    return withStatus(FULFILLING);
   }
 
   public OrderDataBuilder withOrderedStatus() {